requirements.txt
WTForms~=3.0.1
Werkzeug~=2.2.2
Flask~=2.2.2
alembic~=1.9.1
SQLAlchemy~=1.4.46

python-dotenv~=0.21.0

PyJWT~=2.6.0

langdetect~=1.0.9
